requirements.txt
numpy==2.3.1
pandas==2.3.0
scikit-learn==1.6.1
matplotlib==3.10.3
requests==2.32.3
PyYAML==6.0.2
selenium==4.33.0
tqdm==4.67.1
beautifulsoup4==4.13.4
sentence-transformers==4.1.0
joblib==1.5.0
sacrebleu==2.5.1
pytorch-ignite==0.5.2
seaborn==0.13.2
torch==2.7.0
zss==1.2.0
scipy==1.15.3
statsmodels==0.14.2
ollama==0.4.8
openai==1.78.1